+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -51,6 +51,7 @@ import io.getstream.chat.android.client.helpers.QueryChannelsPostponeHelper
 import io.getstream.chat.android.client.logger.ChatLogLevel
 import io.getstream.chat.android.client.logger.ChatLogger
 import io.getstream.chat.android.client.logger.ChatLoggerHandler
+import io.getstream.chat.android.client.models.AppSettings
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.BannedUser
 import io.getstream.chat.android.client.models.BannedUsersSort
@@ -303,6 +304,13 @@ public class ChatClient internal constructor(
         tokenManager.setTokenProvider(tokenProvider)
         warmUp()
     }
+
+    /**
+     * Get the current settings of the app. Check [AppSettings].
+     *
+     * @return [AppSettings] the settings of the app.
+     */
+    public fun appSettings(): Call<AppSettings> = api.appSettings()
 
     /**
      * Initializes [ChatClient] for a specific user. The [tokenProvider] implementation is used
@@ -1065,6 +1073,7 @@ public class ChatClient internal constructor(
             .doOnResult(scope) { result ->
                 plugins.forEach { it.onQueryChannelsResult(result, request) }
             }
+            .precondition { onQueryChannelsPrecondition(request) }
 
     @CheckResult
     public fun deleteChannel(channelType: String, channelId: String): Call<Channel> {
@@ -1099,13 +1108,22 @@ public class ChatClient internal constructor(
      *
      * @param channelType The channel type. ie messaging.
      * @param channelId The channel id. ie 123.
+     * @param systemMessage The system message that will be shown in the channel.
      *
      * @return Executable async [Call] which completes with [Result] having data equal to the truncated channel
      * if the channel was successfully truncated.
      */
     @CheckResult
-    public fun truncateChannel(channelType: String, channelId: String): Call<Channel> {
-        return api.truncateChannel(channelType, channelId)
+    public fun truncateChannel(
+        channelType: String,
+        channelId: String,
+        systemMessage: Message? = null,
+    ): Call<Channel> {
+        return api.truncateChannel(
+            channelType = channelType,
+            channelId = channelId,
+            systemMessage = systemMessage
+        )
     }
 
     @CheckResult
@@ -1244,9 +1262,16 @@ public class ChatClient internal constructor(
         return api.markAllRead()
     }
 
+    /**
+     * Marks the specified channel as read.
+     *
+     * @param channelType Type of the channel.
+     * @param channelId Id of the channel.
+     */
     @CheckResult
     public fun markRead(channelType: String, channelId: String): Call<Unit> {
         return api.markRead(channelType, channelId)
+            .precondition { onChannelMarkReadPrecondition(channelType, channelId) }
     }
 
     @CheckResult
@@ -1304,6 +1329,13 @@ public class ChatClient internal constructor(
         )
     }
 
+    /**
+     * Method to remove members of a given channel.
+     *
+     * @param channelType The channel type. ie messaging.
+     * @param channelId The channel id. ie 123.
+     * @param members The list of the members to be removed.
+     */
     @CheckResult
     public fun removeMembers(
         channelType: String,