+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/ChannelMapping.kt
@@ -29,5 +29,6 @@ internal fun DownstreamChannelDto.toDomain(): Channel =
         team = team,
         cooldown = cooldown,
         pinnedMessages = pinned_messages.map(DownstreamMessageDto::toDomain),
+        ownCapabilities = own_capabilities.toSet(),
         extraData = extraData.toMutableMap(),
     )