+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/parser2/MoshiChatParser.kt
@@ -8,6 +8,7 @@ import io.getstream.chat.android.client.api2.mapping.toDto
 import io.getstream.chat.android.client.api2.model.dto.ChatEventDto
 import io.getstream.chat.android.client.api2.model.dto.UpstreamConnectedEventDto
 import io.getstream.chat.android.client.api2.model.response.SocketErrorResponse
+import io.getstream.chat.android.client.events.ChannelTruncatedEvent
 import io.getstream.chat.android.client.events.ChannelUpdatedByUserEvent
 import io.getstream.chat.android.client.events.ChannelUpdatedEvent
 import io.getstream.chat.android.client.events.ChatEvent
@@ -147,6 +148,7 @@ internal class MoshiChatParser : ChatParser {
             is ReactionUpdateEvent -> message.enrichWithCid(cid)
             is ReactionDeletedEvent -> message.enrichWithCid(cid)
             is ChannelUpdatedEvent -> message?.enrichWithCid(cid)
+            is ChannelTruncatedEvent -> message?.enrichWithCid(cid)
             is ChannelUpdatedByUserEvent -> message?.enrichWithCid(cid)
             is NotificationMessageNewEvent -> message.enrichWithCid(cid)
             else -> { /* Do nothing */