+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/reactions/edit/internal/EditReactionsBubbleDrawer.kt
@@ -1,12 +1,17 @@
 package io.getstream.chat.android.ui.message.list.reactions.edit.internal
 
+import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Paint
 import io.getstream.chat.android.ui.common.extensions.internal.dpToPx
 import io.getstream.chat.android.ui.message.list.reactions.edit.EditReactionsViewStyle
+import io.getstream.chat.android.ui.utils.isRtlLayout
 
 private const val LARGE_TAIL_BUBBLE_OFFSET_CORRECTION_DP = 2
 
+/**
+ * Draws the edit reactions bubble.
+ */
 internal class EditReactionsBubbleDrawer(
     private val editReactionsViewStyle: EditReactionsViewStyle,
 ) {
@@ -26,7 +31,18 @@ internal class EditReactionsBubbleDrawer(
     private var isMyMessage: Boolean = false
     private var isSingleReaction: Boolean = false
 
+    /**
+     * Draws the bubble of reactions choosing the correct direction.
+     *
+     * @param context [Context].
+     * @param canvas [Canvas].
+     * @param bubbleWidth The width of the bubble. This should be at least bigger than all the columns of reactions.
+     * @param bubbleHeight The height of the bubble. This should be at least bigger than all the lines of reactions.
+     * @param isMyMessage Whether this is the message of the current user or not.
+     * @param isSingleReaction Whether there's only a single reaction of there are multiple reactions.
+     */
     fun drawReactionsBubble(
+        context: Context,
         canvas: Canvas,
         bubbleWidth: Int,
         bubbleHeight: Int,
@@ -39,11 +55,19 @@ internal class EditReactionsBubbleDrawer(
         this.isSingleReaction = isSingleReaction
 
         val bubblePaint = if (isMyMessage) bubblePaintMine else bubblePaintTheirs
+        val isRtl = context.isRtlLayout
+
         drawBubbleRoundRect(canvas, bubblePaint)
-        drawLargeTailBubble(canvas, bubblePaint)
-        drawSmallTailBubble(canvas, bubblePaint)
+        drawLargeTailBubble(canvas, bubblePaint, isRtl)
+        drawSmallTailBubble(canvas, bubblePaint, isRtl)
     }
 
+    /**
+     * Draws the main bubble of reactions. The one that involves the reactions.
+     *
+     * @param canvas [Canvas].
+     * @param paint [Paint].
+     */
     private fun drawBubbleRoundRect(canvas: Canvas, paint: Paint) {
         canvas.drawRoundRect(
             0f,
@@ -56,9 +80,16 @@ internal class EditReactionsBubbleDrawer(
         )
     }
 
-    private fun drawLargeTailBubble(canvas: Canvas, paint: Paint) {
-        val offset = editReactionsViewStyle.largeTailBubbleOffset.toFloat().let {
-            if (isMyMessage) it else -it
+    /**
+     * Draws the large tail bubble, normally right bellow the main bubble.
+     *
+     * @param canvas [Canvas].
+     * @param paint [Paint].
+     * @param isRtl If the bubble should be drawn with inverted direction.
+     */
+    private fun drawLargeTailBubble(canvas: Canvas, paint: Paint, isRtl: Boolean) {
+        val offset = editReactionsViewStyle.largeTailBubbleOffset.toFloat().let { bubbleOffset ->
+            parseOffset(isRtl, isMyMessage, bubbleOffset)
         }
         canvas.drawCircle(
             (bubbleWidth / 2).toFloat() + offset,
@@ -68,10 +99,39 @@ internal class EditReactionsBubbleDrawer(
         )
     }
 
-    private fun drawSmallTailBubble(canvas: Canvas, paint: Paint) {
-        val offset = editReactionsViewStyle.smallTailBubbleOffset.toFloat().let {
-            if (isMyMessage) it else -it
+    /**
+     * Parses the offset of the bubble.
+     *
+     * @param isRtl If the bubble offset should be drawn with inverted direction.
+     * @param isMyMessage Whether this is the message of the current user or not.
+     * @param offset The offset of the bubble.
+     */
+    private fun parseOffset(isRtl: Boolean, isMyMessage: Boolean, offset: Float): Float {
+        return when {
+            isMyMessage && !isRtl -> offset
+
+            isMyMessage && isRtl -> -offset
+
+            !isMyMessage && !isRtl -> -offset
+
+            !isMyMessage && isRtl -> offset
+
+            else -> offset
         }
+    }
+
+    /**
+     * Draws the large tail bubble, normally right bellow the large tail bubble.
+     *
+     * @param canvas [Canvas].
+     * @param paint [Paint].
+     * @param isRtl If the bubble should be drawn with inverted direction.
+     */
+    private fun drawSmallTailBubble(canvas: Canvas, paint: Paint, isRtl: Boolean) {
+        val offset = editReactionsViewStyle.smallTailBubbleOffset.toFloat().let { bubbleOffset ->
+            parseOffset(isRtl, isMyMessage, bubbleOffset)
+        }
+
         canvas.drawCircle(
             bubbleWidth / 2 + offset,
             largeTailBubbleInitialPosition() +
@@ -82,6 +142,9 @@ internal class EditReactionsBubbleDrawer(
         )
     }
 
+    /**
+     * The large tail bubble initial position.
+     */
     private fun largeTailBubbleInitialPosition(): Float {
         return bubbleHeight.toFloat() - LARGE_TAIL_BUBBLE_OFFSET_CORRECTION_DP.dpToPx()
     }